Add files via upload
Task 1: Web scraping to get business intelligence
Customers who book a flight with BA will interact with the BA brand on many different levels. Any company, including BA, that wants to succeed must be able to comprehend its customers' thoughts, desires, and feedback.

The initial step is to gather and examine data from a third-party source, scrape customer reviews, and then analyze the data to present any insights found.

Skytrax was used to gather information on customer opinions of British Airways.

The following findings are encapsulated in the single-slide presentation.